Improved Text Galaxy and Added option to choose percentage needed to win council vote.
- Fixed possible crash while changing the number of companion world.
  - Added base galaxy size for Dynamic Galaxy with a general formula, valid for all galaxy size.
  - Text shaped Galaxy: Added option for "Fusion" font.
  - Text shaped Galaxy: Added selection from file Galaxy.txt
   - Fixed international font problems.
  - Added option to choose percentage needed to win council vote.

+++ src/rotp/model/empires/CustomRaceDefinitions.java
@@ -1670,7 +1670,7 @@ public class CustomRaceDefinitions  {
 			private static final float	c2 = 9.56463523f;
 			private static final float	c3 = -4.365405984f;
 			private static final float	c4 = 0.824090347f;
-			private static final int	baseCostDefault = 100;
+			private static final float	baseCostDefault = 100f;
 			private static final float	norm = 100f;
 
 			private SettingResearch(String nameLangLabel) {
@@ -1861,7 +1861,7 @@ public class CustomRaceDefinitions  {
 			private static final float	c3 = -2.37443919f;
 			private static final float	c4 = -0.62901149f;
 			private static final float	c5 = 0.576847553f;
-			private static final int	baseCostDefault = 50;
+			private static final float	baseCostDefault = 50f;
 			private static final float	norm = 50f;
 
 			private SettingDiscovery(String nameLangLabel) {
